backend/requirements.txt
amqp==5.2.0
annotated-types==0.6.0
anyio==4.4.0
appdirs==1.4.4
APScheduler==3.10.4
asttokens==2.4.0
awscli==1.29.62
awsebcli==3.20.10
backcall==0.2.0
billiard==4.2.0
blessed==1.20.0
blinker==1.6.3
boto3==1.28.62
botocore==1.31.62
CacheControl==0.13.1
cachetools==5.3.2
celery==5.4.0
cement==2.8.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.3
cryptography==41.0.7
decorator==4.4.2
docopt==0.6.2
docutils==0.16
elevenlabs==1.2.2
exceptiongroup==1.2.0
executing==2.0.0
filelock==3.9.0
firebase-admin==6.3.0
Flask==3.0.0
Flask-Cors==4.0.0
fs==2.4.16
fsspec==2023.4.0
google-api-core==2.15.0
google-api-python-client==2.113.0
google-auth==2.26.2
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-status==1.60.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.4
imageio==2.31.5
imageio-ffmpeg==0.4.9
importlib-metadata==7.0.1
iniconfig==2.0.0
ipython==8.16.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
kombu==5.3.7
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mpmath==1.3.0
msgpack==1.0.7
networkx==3.0
numpy==1.26.0
packaging==23.2
parso==0.8.3
pathspec==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
pipreqs==0.4.13
pluggy==1.3.0
proglog==0.1.10
prompt-toolkit==3.0.39
proto-plus==1.23.0
protobuf==4.25.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycurl==7.45.3
pydantic==2.4.2
pydantic_core==2.10.1
pydub==0.25.1
Pygments==2.16.1
PyJWT==2.8.0
pyparsing==3.1.1
pytest==7.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-mock==1.11.0
rsa==4.7.2
s3transfer==0.7.0
semantic-version==2.8.5
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
sympy==1.12
termcolor==1.1.0
tomli==2.0.1
tqdm==4.66.1
traitlets==5.11.2
typing_extensions==4.8.0
tzdata==2024.1
tzlocal==5.1
uritemplate==4.1.1
urllib3==1.26.17
vine==5.1.0
wcwidth==0.1.9
websockets==11.0.3
Werkzeug==3.0.0
yarg==0.1.9
zipp==3.17.0
